+++ src/main/java/cl/duoc/wsofertas/entities/Rubro.java
@@ -7,6 +7,7 @@ package cl.duoc.wsofertas.entities;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 import javax.persistence.Basic;
@@ -56,7 +57,7 @@ public class Rubro implements Serializable {
     @Column(name = "DESCRIPCION")
     private String descripcion;
     @Column(name = "ISACTIVO")
-    private Short isactivo;
+    private BigInteger isactivo;
     @Column(name = "FECHACREACION")
     @Temporal(TemporalType.TIMESTAMP)
     private Date fechacreacion;
@@ -99,11 +100,11 @@ public class Rubro implements Serializable {
         this.descripcion = descripcion;
     }
 
-    public Short getIsactivo() {
+    public BigInteger getIsactivo() {
         return isactivo;
     }
 
-    public void setIsactivo(Short isactivo) {
+    public void setIsactivo(BigInteger isactivo) {
         this.isactivo = isactivo;
     }
